+++ julian.cpp
@@ -98,10 +98,10 @@ void lab2::Julian::jdn_to_julian_date(const long jdn,
     unsigned int s = 153;
     unsigned int w = 2;
 
-    unsigned int f = jdn + j;
-    unsigned int e = (r * f) + v;
-    unsigned int g = (e % p) / r;
-    unsigned int h = (u * g) + w;
+    unsigned long f = jdn + j;
+    unsigned long e = (r * f) + v;
+    unsigned long g = (e % p) / r;
+    unsigned long h = (u * g) + w;
     day = ((h % s) / u) + 1;
     month = (((h / s) + m) % n) + 1;
     year = (e / p) - y + ((n + m - month) / n);
